resource/v1: define resource & relationship types
A Resource is a typed node on the Infrastructure graph and represents an
instance of a cloud resource or workload.

A Relationship is a directed edge that represents a relationship between
two Resources. It is defined like a RDF triplet that contains a
predicate, which defines what type of relationship it is, and a subject
and object for the predicate. There can be multiple relationships with
the same predicate type attached to the same resource.

Many relationships between Resources in the resource graph are bidirectional.
Bidirectional edges are modeled as two separate Relationship instances.
When Relationships are updated in the graph the inverse Relationships
should be kept up to date.

Signed-off-by: Hamzah Qudsi <hamzah@antimetal.com>