resource/v1: refactor: harden type definitions
Harden the API structure for resource objects. The API introduced by
this commit provides a generic structure to represent objects in the
infrastructure resource graph whether the be Resource nodes on the graph
or Relationship types.

The API is defined as follows:

An "Object" is a typed entity in the graph. An Object can either by a
graph node or edge. Every object self-describes its type known as a
"TypeDescriptor", which determines what field attributes the object has.
The TypeDescriptor contains the following:

 - kind: The graph entity type. It denotes whether the object is a
   "Resource" (graph node) or a "Relationship" (graph edge)

 - type: the full protobuf type url of the object schema (e.g.
   aws.ec2.v1.Instance, k8s.io.api.apps.v1.Deployment)

So an Object that has the TypeDescriptor {kind: "Resource", type:
aws.ec2.v1.Instance} represents an AWS EC2 instance that is a node on
the graph whose schema is defined by the aws.ec2.v1.Instance protobuf
type.

There are two defined Object Kinds:

- Resource: Represents a cloud resource (e.g. AWS EC2 instance,
  Kubernetes pod, etc.) A Resource's TypeDescriptor type defines its
  spec schema.

- "Relationship": A directed edge between two Resources. It is defined
  like a triplet that contains a predicate, which defines what type of
  relationship  it is, and a subject and object for the predicate.
  A Relationships's TypeDescriptor type defines what the predicate
  type is. There can be multiple associations with the same predicate
  type attached to the same resource.

With this new API structure the delta types in package resource.v1 are
moved to service.resource.v1 since they are now very specific to the
intake service.

Signed-off-by: Hamzah Qudsi <hamzah@antimetal.com>