refactor: make Collectors string instead of enum.
Number of types of collectors are changing rapidly so it's annoying to
keep the enum updated. The Agent Management Server will store which
collectors an agent supports which can be used to specify which
collectors can be configured.

This also allows for easier dynamic updates when an Agent supports a new
collector.

AgentManagementService hasn't been implemented yet so we can make
breaking changes.

Signed-off-by: Hamzah Qudsi <hamzah@antimetal.com>